requirements.txt
khl.py
aiosqlite
httpx
bs4
matplotlib